Make source filenames match module names on case-sensitive filesystems.
Prior to this commit, the `Vega` module is in a file `src/vega.elm`.  On
case-sensitive filesystems (e.g. Linux) this causes the build to fail as
follows:

```
I cannot find module 'Vega'.

This module is demanded in elm-package.json.

Potential problems could be:
  * Misspelled the module name
  * Need to add a source directory or new dependency to elm-package.json
```

Renaming `src/vega.elm` to `src/Vega.elm` (and likewise for the other source
files) resolves this; that is what this commit does.